server/requirements.txt
PySensors==0.0.3
argparse==1.2.1
pyserial==2.7
